+++ html/src/playn/html/HtmlGraphicsCanvas.java
@@ -59,7 +59,7 @@ class HtmlGraphicsCanvas extends HtmlGraphics {
   }
 
   @Override
-  public SurfaceLayer createSurfaceLayer(int width, int height) {
+  public SurfaceLayer createSurfaceLayer(float width, float height) {
     return new HtmlSurfaceLayerCanvas(width, height);
   }
